update ALI.py run.py dispatcher.py

+++ dpgen/dispatcher/Dispatcher.py
@@ -48,6 +48,8 @@ class Dispatcher(object):
         elif context_type == 'ssh':
             self.session = SSHSession(remote_profile)
             self.context = SSHContext
+            self.checkContext = None
+            self.checkBatch = None
             self.uuid_names = False
         else :
             raise RuntimeError('unknown context')
@@ -65,6 +67,7 @@ class Dispatcher(object):
             raise RuntimeError('unknown batch ' + batch_type)
 
 
+        # when this function was called, the instance has been already created
     def run_jobs(self,
                  resources,
                  command,
@@ -76,7 +79,8 @@ class Dispatcher(object):
                  backward_task_files,
                  forward_task_deference = True,
                  outlog = 'log',
-                 errlog = 'err') :
+                 errlog = 'err',
+                 machine = None) :
         # task_chunks = [
         #     [os.path.basename(j) for j in tasks[i:i + group_size]] \
         #     for i in range(0, len(tasks), group_size)
@@ -96,14 +100,15 @@ class Dispatcher(object):
                 chunk_sha1 = sha1(task_chunks_[ii].encode('utf-8')).hexdigest() 
                 # if hash in map, recover job, else start a new job
                 if chunk_sha1 in path_map:
-                    # job_uuid = path_map[chunk_sha1][1].split('/')[-1]
-                    job_uuid = path_map[chunk_sha1][2]
+                    job_uuid = path_map[chunk_sha1][1].split('/')[-1]
                     dlog.debug("load uuid %s for chunk %s" % (job_uuid, task_chunks_[ii]))
                 else:
                     job_uuid = None
                 # communication context, bach system
                 context = self.context(work_path, self.session, job_uuid)
+                self.checkContext = context
                 batch = self.batch(context, uuid_names = self.uuid_names)
+                self.chechBatch = batch
                 rjob = {'context':context, 'batch':batch}
                 # upload files
                 if not rjob['context'].check_file_exists('tag_upload'):
@@ -117,45 +122,51 @@ class Dispatcher(object):
                 # submit new or recover old submission
                 if job_uuid is None:
                     rjob['batch'].submit(chunk, command, res = resources, outlog=outlog, errlog=errlog)
-                    job_uuid = rjob['context'].job_uuid
-                    dlog.debug('assigned uudi %s for %s ' % (job_uuid, task_chunks_[ii]))
-                    dlog.info('new submission of %s' % job_uuid)
+                    dlog.debug('assigned uudi %s for %s ' % (rjob['context'].job_uuid, task_chunks_[ii]))
+                    dlog.info('new submission of %s' % rjob['context'].job_uuid)
                 else:
                     rjob['batch'].submit(chunk, command, res = resources, outlog=outlog, errlog=errlog, restart = True)
                     dlog.info('restart from old submission %s ' % job_uuid)
                 # record job and its hash
                 job_list.append(rjob)
-                path_map[chunk_sha1] = [context.local_root, context.remote_root, job_uuid]
+                path_map[chunk_sha1] = [context.local_root,context.remote_root]
             else :
                 # finished job, append a None to list
                 job_list.append(None)
-        _pmap.dump(path_map)
 
-        assert(len(job_list) == len(task_chunks))
-        fcount = [0]*len(job_list)
-        while not all(job_fin) :
-            dlog.debug('checking jobs')
-            for idx,rjob in enumerate(job_list) :
-                if not job_fin[idx] :
-                    status = rjob['batch'].check_status()
-                    job_uuid = rjob['context'].job_uuid
-                    if status == JobStatus.terminated :
-                        fcount[idx] += 1
-                        if fcount[idx] > 3:
-                            raise RuntimeError('Job %s failed for more than 3 times' % job_uuid)
-                        dlog.info('job %s terminated, submit again'% job_uuid)
-                        dlog.debug('try %s times for %s'% (fcount[idx], job_uuid))
-                        rjob['batch'].submit(task_chunks[idx], command, res = resources, outlog=outlog, errlog=errlog,restart=True)
-                    elif status == JobStatus.finished :
-                        dlog.info('job %s finished' % job_uuid)
-                        rjob['context'].download(task_chunks[idx], backward_task_files)
-                        rjob['context'].clean()
-                        job_fin[idx] = True
-                        _fr.write_record(job_fin)
-            time.sleep(10)
-        # delete path map file when job finish
-        _pmap.delete()
-
+        if machine == 'ali':
+            status = rjob['batch'].check_status()
+            job_uuid = rjob['context'].job_uuid
+            return status
+        else:
+            _pmap.dump(path_map)
+            assert(len(job_list) == len(task_chunks))
+            fcount = [0]*len(job_list)
+            while not all(job_fin) :
+                dlog.debug('checking jobs')
+                for idx,rjob in enumerate(job_list) :
+                    if not job_fin[idx] :
+                        status = rjob['batch'].check_status()
+                        job_uuid = rjob['context'].job_uuid
+                        if status == JobStatus.terminated :
+                            fcount[idx] += 1
+                            if fcount[idx] > 3:
+                                raise RuntimeError('Job %s failed for more than 3 times' % job_uuid)
+                            dlog.info('job %s terminated, submit again'% job_uuid)
+                            dlog.debug('try %s times for %s'% (fcount[idx], job_uuid))
+                            rjob['batch'].submit(task_chunks[idx], command, res = resources, outlog=outlog, errlog=errlog,restart=True)
+                        elif status == JobStatus.finished :
+                            dlog.info('job %s finished' % job_uuid)
+                            rjob['context'].download(task_chunks[idx], backward_task_files)
+                            rjob['context'].clean()
+                            job_fin[idx] = True
+                            _fr.write_record(job_fin)
+                time.sleep(10)
+            # delete path map file when job finish
+            _pmap.delete()
+    # this function will be called after run_jobs, so self.chexkContext has already initialized.
+    def dis_check_status(self):
+        return self.checkBatch.check_status()
 
 class FinRecord(object):
     def __init__ (self, path, njobs, fname = 'fin.record'):
